Initial commit for FMG proxy security profiles
