+++ test/units/modules/network/fortimanager/fixtures/test_fmgr_secprof_proxy.json
@@ -1,0 +1,149 @@
+{
+   "fmgr_web_proxy_profile_addsetdelete": [
+      {
+         "raw_response": {
+            "status": {
+               "message": "Object does not exist", 
+               "code": -3
+            }, 
+            "url": "/pm/config/adom/root/obj/web-proxy/profile/Ansible_Web_Filter_Profile"
+         }, 
+         "paramgram_used": {
+            "header-via-request": null, 
+            "name": "Ansible_Web_Filter_Profile", 
+            "adom": "root", 
+            "log-header-change": null, 
+            "header-front-end-https": null, 
+            "headers": {
+               "action": null, 
+               "content": null, 
+               "name": null
+            }, 
+            "mode": "delete", 
+            "header-via-response": null, 
+            "header-x-authenticated-user": null, 
+            "strip-encoding": null, 
+            "header-x-forwarded-for": null, 
+            "header-x-authenticated-groups": null, 
+            "header-client-ip": null
+         }, 
+         "post_method": "delete"
+      }, 
+      {
+         "paramgram_used": {
+            "header-via-request": null, 
+            "name": "Ansible_Web_Filter_Profile", 
+            "header-front-end-https": null, 
+            "log-header-change": null, 
+            "adom": "root", 
+            "headers": {
+               "action": null, 
+               "content": null, 
+               "name": null
+            }, 
+            "mode": "delete", 
+            "header-via-response": null, 
+            "header-x-authenticated-user": null, 
+            "strip-encoding": null, 
+            "header-x-forwarded-for": null, 
+            "header-x-authenticated-groups": null, 
+            "header-client-ip": null
+         }, 
+         "raw_response": {
+            "status": {
+               "message": "Object does not exist", 
+               "code": -3
+            }, 
+            "url": "/pm/config/adom/root/obj/web-proxy/profile/Ansible_Web_Filter_Profile"
+         }, 
+         "post_method": "delete"
+      }, 
+      {
+         "raw_response": {
+            "status": {
+               "message": "OK", 
+               "code": 0
+            }, 
+            "url": "/pm/config/adom/root/obj/web-proxy/profile"
+         }, 
+         "paramgram_used": {
+            "header-via-request": null, 
+            "name": "Ansible_Web_Proxy_Profile", 
+            "adom": "root", 
+            "log-header-change": null, 
+            "header-front-end-https": null, 
+            "headers": {
+               "action": null, 
+               "content": null, 
+               "name": null
+            }, 
+            "mode": "set", 
+            "header-via-response": null, 
+            "header-x-authenticated-user": null, 
+            "strip-encoding": null, 
+            "header-x-forwarded-for": null, 
+            "header-x-authenticated-groups": null, 
+            "header-client-ip": null
+         }, 
+         "post_method": "set"
+      }, 
+      {
+         "paramgram_used": {
+            "header-via-request": null, 
+            "name": "Ansible_Web_Proxy_Profile", 
+            "header-front-end-https": null, 
+            "log-header-change": null, 
+            "adom": "root", 
+            "headers": {
+               "action": null, 
+               "content": null, 
+               "name": null
+            }, 
+            "mode": "delete", 
+            "header-via-response": null, 
+            "header-x-authenticated-user": null, 
+            "strip-encoding": null, 
+            "header-x-forwarded-for": null, 
+            "header-x-authenticated-groups": null, 
+            "header-client-ip": null
+         }, 
+         "raw_response": {
+            "status": {
+               "message": "OK", 
+               "code": 0
+            }, 
+            "url": "/pm/config/adom/root/obj/web-proxy/profile/Ansible_Web_Proxy_Profile"
+         }, 
+         "post_method": "delete"
+      }, 
+      {
+         "raw_response": {
+            "status": {
+               "message": "OK", 
+               "code": 0
+            }, 
+            "url": "/pm/config/adom/root/obj/web-proxy/profile"
+         }, 
+         "paramgram_used": {
+            "header-via-request": "remove", 
+            "header-client-ip": "pass", 
+            "header-front-end-https": "add", 
+            "header-x-authenticated-groups": "add", 
+            "name": "Ansible_Web_Proxy_Profile", 
+            "log-header-change": "enable", 
+            "adom": "root", 
+            "headers": {
+               "action": "add-to-request", 
+               "content": "test", 
+               "name": "test_header"
+            }, 
+            "mode": "set", 
+            "header-via-response": "pass", 
+            "header-x-authenticated-user": "remove", 
+            "strip-encoding": "enable", 
+            "header-x-forwarded-for": "pass"
+         }, 
+         "post_method": "set"
+      }
+   ]
+}
