--- test/units/modules/network/fortimanager/fixtures/test_fmgr_secprof_web.json
@@ -1,214 +1,0 @@
-{
-   "fmgr_webfilter_profile_addsetdelete": [
-      {
-         "paramgram_used": {
-            "comment": null, 
-            "web-extended-all-action-log": null, 
-            "url-extraction": {
-               "status": null, 
-               "redirect-url": null, 
-               "server-fqdn": null, 
-               "redirect-header": null, 
-               "redirect-no-content": null
-            }, 
-            "ftgd-wf": {
-               "filters": {
-                  "category": null, 
-                  "log": null, 
-                  "override-replacemsg": null, 
-                  "warning-duration-type": null, 
-                  "warn-duration": null, 
-                  "auth-usr-grp": null, 
-                  "action": null, 
-                  "warning-prompt": null
-               }, 
-               "ovrd": null, 
-               "rate-image-urls": null, 
-               "quota": {
-                  "category": null, 
-                  "value": null, 
-                  "override-replacemsg": null, 
-                  "duration": null, 
-                  "type": null, 
-                  "unit": null
-               }, 
-               "options": null, 
-               "rate-javascript-urls": null, 
-               "max-quota-timeout": null, 
-               "rate-css-urls": null, 
-               "exempt-quota": null, 
-               "rate-crl-urls": null
-            }, 
-            "log-all-url": null, 
-            "extended-log": null, 
-            "web-filter-cookie-removal-log": null, 
-            "https-replacemsg": null, 
-            "web": {
-               "log-search": null, 
-               "bword-threshold": null, 
-               "bword-table": null, 
-               "whitelist": null, 
-               "youtube-restrict": null, 
-               "safe-search": null, 
-               "blacklist": null, 
-               "keyword-match": null, 
-               "urlfilter-table": null, 
-               "content-header-list": null
-            }, 
-            "wisp": null, 
-            "web-filter-vbs-log": null, 
-            "youtube-channel-filter": {
-               "comment": null, 
-               "channel-id": null
-            }, 
-            "override": {
-               "profile": null, 
-               "ovrd-scope": null, 
-               "ovrd-dur-mode": null, 
-               "profile-attribute": null, 
-               "ovrd-dur": null, 
-               "profile-type": null, 
-               "ovrd-user-group": null, 
-               "ovrd-cookie": null
-            }, 
-            "ovrd-perm": null, 
-            "mode": "delete", 
-            "web-content-log": null, 
-            "web-invalid-domain-log": null, 
-            "adom": "root", 
-            "web-filter-referer-log": null, 
-            "inspection-mode": null, 
-            "post-action": null, 
-            "name": "Ansible_Web_Filter_Profile", 
-            "web-filter-command-block-log": null, 
-            "web-filter-applet-log": null, 
-            "web-ftgd-err-log": null, 
-            "replacemsg-group": null, 
-            "web-filter-jscript-log": null, 
-            "web-filter-activex-log": null, 
-            "web-filter-js-log": null, 
-            "web-ftgd-quota-usage": null, 
-            "web-filter-unknown-log": null, 
-            "web-filter-cookie-log": null, 
-            "youtube-channel-status": null, 
-            "web-url-log": null, 
-            "options": null, 
-            "wisp-servers": null, 
-            "wisp-algorithm": null
-         }, 
-         "raw_response": {
-            "status": {
-               "message": "OK", 
-               "code": 0
-            }, 
-            "url": "/pm/config/adom/root/obj/webfilter/profile/Ansible_Web_Filter_Profile"
-         }, 
-         "post_method": "delete"
-      }, 
-      {
-         "raw_response": {
-            "status": {
-               "message": "OK", 
-               "code": 0
-            }, 
-            "url": "/pm/config/adom/root/obj/webfilter/profile"
-         }, 
-         "paramgram_used": {
-            "comment": "Created by Ansible Module TEST", 
-            "web-filter-command-block-log": "enable", 
-            "web-invalid-domain-log": "enable", 
-            "web-extended-all-action-log": "enable", 
-            "adom": "root", 
-            "ftgd-wf": {
-               "rate-javascript-urls": null, 
-               "quota": {
-                  "category": null, 
-                  "value": null, 
-                  "override-replacemsg": null, 
-                  "duration": null, 
-                  "type": null, 
-                  "unit": null
-               }, 
-               "rate-image-urls": null, 
-               "filters": {
-                  "category": null, 
-                  "auth-usr-grp": null, 
-                  "log": null, 
-                  "warning-prompt": null, 
-                  "override-replacemsg": null, 
-                  "action": null, 
-                  "warn-duration": null, 
-                  "warning-duration-type": null
-               }, 
-               "rate-css-urls": null, 
-               "ovrd": null, 
-               "exempt-quota": null, 
-               "max-quota-timeout": null, 
-               "rate-crl-urls": null, 
-               "options": null
-            }, 
-            "web-content-log": "enable", 
-            "web-filter-referer-log": "enable", 
-            "log-all-url": "enable", 
-            "extended-log": "enable", 
-            "inspection-mode": "proxy", 
-            "web-filter-cookie-removal-log": "enable", 
-            "post-action": "block", 
-            "web-filter-activex-log": "enable", 
-            "web-filter-cookie-log": "enable", 
-            "web": {
-               "blacklist": null, 
-               "log-search": null, 
-               "keyword-match": null, 
-               "urlfilter-table": null, 
-               "bword-table": null, 
-               "safe-search": null, 
-               "whitelist": null, 
-               "content-header-list": null, 
-               "youtube-restrict": null, 
-               "bword-threshold": null
-            }, 
-            "web-filter-applet-log": "enable", 
-            "web-ftgd-err-log": "enable", 
-            "replacemsg-group": null, 
-            "web-filter-jscript-log": "enable", 
-            "web-ftgd-quota-usage": "enable", 
-            "url-extraction": {
-               "status": null, 
-               "server-fqdn": null, 
-               "redirect-url": null, 
-               "redirect-header": null, 
-               "redirect-no-content": null
-            }, 
-            "web-filter-js-log": "enable", 
-            "youtube-channel-filter": {
-               "comment": null, 
-               "channel-id": null
-            }, 
-            "name": "Ansible_Web_Filter_Profile", 
-            "wisp": "enable", 
-            "web-filter-vbs-log": "enable", 
-            "web-filter-unknown-log": "enable", 
-            "mode": "set", 
-            "youtube-channel-status": "blacklist", 
-            "override": {
-               "profile": null, 
-               "ovrd-user-group": null, 
-               "ovrd-scope": null, 
-               "ovrd-cookie": null, 
-               "ovrd-dur-mode": null, 
-               "profile-attribute": null, 
-               "ovrd-dur": null, 
-               "profile-type": null
-            }, 
-            "web-url-log": "enable", 
-            "ovrd-perm": "bannedword-override", 
-            "https-replacemsg": null, 
-            "options": "js", 
-            "wisp-servers": null, 
-            "wisp-algorithm": "auto-learning"
-         }, 
-         "post_method": "set"
-      }
-   ]
-}
